agents/writer/requirements.txt
strands-agents[a2a]
bedrock-agentcore[a2a]
boto3